pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulumi-lagoon"
version = "0.1.1"
description = "A Pulumi dynamic provider for managing Lagoon resources"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Greg Chaix", email = "greg@tag1consulting.com"}
]
maintainers = [
    {name = "Tag1 Consulting", email = "info@tag1consulting.com"}
]
keywords = [
    "pulumi",
    "lagoon",
    "infrastructure-as-code",
    "iac",
    "devops",
    "kubernetes",
    "amazee",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "pulumi>=3.0.0,<4.0.0",
    "requests>=2.28.0,<3.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/tag1consulting/pulumi-lagoon-provider"
Documentation = "https://github.com/tag1consulting/pulumi-lagoon-provider#readme"
Repository = "https://github.com/tag1consulting/pulumi-lagoon-provider.git"
Issues = "https://github.com/tag1consulting/pulumi-lagoon-provider/issues"
Changelog = "https://github.com/tag1consulting/pulumi-lagoon-provider/blob/main/RELEASE_NOTES.md"

[tool.setuptools]
packages = ["pulumi_lagoon"]

[tool.setuptools.package-data]
pulumi_lagoon = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise-without-from-inside-except (existing code pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["pulumi_lagoon"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["pulumi_lagoon"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
